Adjust git user identity in publish action
The event payload's "pusher" doesn't reflect the git commits pushed;
it's just the generic username and default email address for the GitHub
account.  This is not as desirable, so make tags as a service identity
instead.  There does appear to be actual git commit information in the
.commits[].author data of the payload, but probably better not to
completely impersonate anyway.