scan-irb: Translate result template to Tigrinya
This requires using locally saved fonts. The only way we could currently
figure out how to access the locally saved fonts in a Docker-compatible
(e.g. production) way was to hardcode the path relative to the workdir
of the Docker container. This broke the ability to run LaTeX outside of
a Docker container in development.