scan-irb: Drop study-specific logos
Drop the SCAN and UW logos from the results PDF because they are not
required. Lab results just need to be specific to NWGC (CLIA name).
Keep the previous commit in case this ends up being not true and we have
to include the separate study logos.

Note: the new logo formatting is updated for the English template only.
A following commit will address the updates for other languages.