scan-irb: Update logo
The previous `logo.pdf` was too small and looked fuzzy when zoomed in to
higher resolutions. Make the logo bigger for better graphics so it looks
crisp at 400x zoom.

Additionally, convert the text elements in the SVG to path elements so
that the image doesn't change depending on what fonts are installed on
someone's computer.