clinical: update new clinical results PDF language and logo
Replace the current header logos with those for BBI, UW Med,
and PHKSC. Make various text updates, replacing IRB research
specific text with text appropriate to the clinical result.
Include additional variable ordering_provider in the example
paramters file.

This template will be used for the Adult Family Homes and
Workplace Outbreak project, which is not IRB, and is currently
available only in English.